Added parameters to from_file and improve names of classifying scripts
